+++ app.go
@@ -93,8 +93,16 @@ func main() {
 			server.ListenAndServe()
 		}
 	}()
-	e := mesos.Subscribe(&config, &framework)
-	e.API = a
-	e.EventLoop()
-	config.RedisClient.Close()
+
+	//	this loop is for resubscribtion purpose
+	ticker := time.NewTicker(2 * time.Second)
+	defer ticker.Stop()
+	for {
+		select {
+		case <-ticker.C:
+			e := mesos.Subscribe(&config, &framework)
+			e.API = a
+			e.EventLoop()
+		}
+	}
 }