+++ changelog.md
@@ -14,6 +14,7 @@
 - ADD: Support for docker comport "pull_policy". Support always (default) and "missing".
 - FIX: Recalculate the HostPorts if the Mesos Task is failed.
 - ADD: Support for docker compose "placement -> constraints -> node.hostname" command.
+- ADD: Resubscription after the connection to mesos master is lost.